* some "inlining" in the PyRep class (UNIX dev's please tell me if it compiles as inline can sometimes be a bitch in GCC). * default sending back PyNone when the client asks for a auth message. * massively added comments to the AlertService class and added some developer treats. * implemented client ping request (it still needs some love as I fake the entire thing). * changed outrole in DoLogin to 64 bits integer (comments like "this should be that..." sucks just do it). * retranslated some "dosent" into "doesn't" * replaced some delete to SafeDelete. * some minor cleaning.

+++ server/ServiceDB.cpp
@@ -37,7 +37,7 @@ ServiceDB::ServiceDB(ServiceDB *existing_db)
 ServiceDB::~ServiceDB() {
 }
 
-bool ServiceDB::DoLogin(const char *login, const char *pass, uint32 &out_accountID, uint32 &out_role) {
+bool ServiceDB::DoLogin(const char *login, const char *pass, uint32 &out_accountID, uint64 &out_role) {
 	DBQueryResult res;
 
 	if(pass[0] == '\0') {
